Internship-project/requirements.txt
ultralytics
opencv-python
pillow
